@@ -56,7 +56,7 @@ def run_add_constraints_example(solver="cbc", nologg=False):
     # ##### creating an oemof solph optimization model, nothing special here ##
     # create an energy system object for the oemof solph nodes
     es = EnergySystem(
-        timeindex=pd.date_range("1/1/2017", periods=5, freq="H"),
+        timeindex=pd.date_range("1/1/2017", periods=5, freq="h"),
         infer_last_interval=False,
     )
     # add some nodes
