@@ -302,4 +302,13 @@ def create_time_index(
         number = round(hoy / interval)
     if start is None:
         start = f"1/1/{year}"
-    return pd.date_range(start, periods=number + 1, freq=f"{interval}h")
+    try:
+        time_index = pd.date_range(
+            start, periods=number + 1, freq=f"{interval}h"
+        )
+    except ValueError:
+        # Pandas <2.2 compatibility
+        time_index = pd.date_range(
+            start, periods=number + 1, freq=f"{interval}H"
+        )
+    return time_index
