@@ -22,20 +22,20 @@ from oemof.solph import EnergySystem
 )
 def test_add_periods():
     """test method _add_periods of energy system"""
-    timeindex = pd.date_range(start="2012-01-01", periods=10000, freq="H")
+    timeindex = pd.date_range(start="2012-01-01", periods=10000, freq="h")
     periods = [
-        pd.date_range(start="2012-01-01", periods=8784, freq="H"),
-        pd.date_range(start="2013-01-01", periods=1217, freq="H"),
+        pd.date_range(start="2012-01-01", periods=8784, freq="h"),
+        pd.date_range(start="2013-01-01", periods=1217, freq="h"),
     ]
     es = EnergySystem(
         timeindex=timeindex, periods=periods, infer_last_interval=True
     )
     assert len(es.periods) == 2
     assert es.periods[0].equals(
-        pd.date_range(start="2012-01-01", periods=8784, freq="H")
+        pd.date_range(start="2012-01-01", periods=8784, freq="h")
     )
     assert es.periods[1].equals(
-        pd.date_range(start="2013-01-01", periods=1217, freq="H")
+        pd.date_range(start="2013-01-01", periods=1217, freq="h")
     )
 
 
@@ -48,9 +48,9 @@ def test_add_periods():
 )
 def test_extract_periods_years():
     """test method _extract_periods_years of energy system"""
-    t_idx_1 = pd.date_range("1/1/2020", periods=3, freq="H").to_series()
-    t_idx_2 = pd.date_range("1/1/2041", periods=3, freq="H").to_series()
-    t_idx_3 = pd.date_range("1/1/2050", periods=3, freq="H").to_series()
+    t_idx_1 = pd.date_range("1/1/2020", periods=3, freq="h").to_series()
+    t_idx_2 = pd.date_range("1/1/2041", periods=3, freq="h").to_series()
+    t_idx_3 = pd.date_range("1/1/2050", periods=3, freq="h").to_series()
     timeindex = pd.concat([t_idx_1, t_idx_2, t_idx_3]).index
     periods = [t_idx_1, t_idx_2, t_idx_3]
     es = EnergySystem(
