@@ -63,7 +63,7 @@ def test_tuples_as_labels_example(
     filename="storage_investment.csv", solver="cbc"
 ):
     logging.info("Initialize the energy system")
-    date_time_index = pd.date_range("1/1/2012", periods=40, freq="H")
+    date_time_index = pd.date_range("1/1/2012", periods=40, freq="h")
 
     energysystem = solph.EnergySystem(
         timeindex=date_time_index,
