@@ -16,28 +16,28 @@ from oemof import solph
 
 def test_energysystem_with_datetimeindex_infer_last_interval():
     """Test EnergySystem with DatetimeIndex (equidistant)"""
-    datetimeindex = pd.date_range("1/1/2012", periods=24, freq="H")
+    datetimeindex = pd.date_range("1/1/2012", periods=24, freq="h")
     es = solph.EnergySystem(timeindex=datetimeindex, infer_last_interval=True)
     assert es.timeincrement[1] == 1.0
     assert es.timeincrement.sum() == 24
 
 
 def test_energysystem_with_datetimeindex():
-    datetimeindex = pd.date_range("1/1/2012", periods=24, freq="H")
+    datetimeindex = pd.date_range("1/1/2012", periods=24, freq="h")
     es = solph.EnergySystem(timeindex=datetimeindex, infer_last_interval=False)
     assert es.timeincrement[1] == 1.0
     assert es.timeincrement.sum() == 23
 
 
 def test_energysystem_interval_inference_warning():
-    datetimeindex = pd.date_range("1/1/2012", periods=24, freq="H")
+    datetimeindex = pd.date_range("1/1/2012", periods=24, freq="h")
     with pytest.warns(FutureWarning):
         _ = solph.EnergySystem(timeindex=datetimeindex)
 
 
 def test_energysystem_with_datetimeindex_non_equidistant_infer_last_interval():
     """Test EnergySystem with DatetimeIndex (non-equidistant)"""
-    dtindex1 = pd.date_range("1/1/2012", periods=24, freq="H")
+    dtindex1 = pd.date_range("1/1/2012", periods=24, freq="h")
     dtindex2 = pd.date_range("1/2/2012", periods=49, freq="30min")
     dtindex = dtindex1.union(dtindex2)
     msg = (
@@ -50,7 +50,7 @@ def test_energysystem_with_datetimeindex_non_equidistant_infer_last_interval():
 
 def test_energysystem_with_datetimeindex_non_equidistant():
     """Test EnergySystem with DatetimeIndex (non-equidistant)"""
-    dtindex1 = pd.date_range("1/1/2012", periods=24, freq="H")
+    dtindex1 = pd.date_range("1/1/2012", periods=24, freq="h")
     dtindex2 = pd.date_range("1/2/2012", periods=49, freq="30min")
     dtindex = dtindex1.union(dtindex2)
     es = solph.EnergySystem(timeindex=dtindex, infer_last_interval=False)
@@ -107,7 +107,7 @@ def test_energysystem_with_numeric_index_non_equidistant():
 
 
 def test_model_timeincrement_with_valid_timeindex():
-    datetimeindex = pd.date_range("1/1/2012", periods=5, freq="H")
+    datetimeindex = pd.date_range("1/1/2012", periods=5, freq="h")
     es = solph.EnergySystem(timeindex=datetimeindex, infer_last_interval=True)
     m = solph._models.BaseModel(es)
     assert es.timeincrement.sum() == 5
@@ -129,7 +129,7 @@ def test_conflicting_time_index():
     )
     with pytest.raises(AttributeError, match=msg):
         solph.EnergySystem(
-            timeindex=pd.date_range("1/1/2012", periods=2, freq="H"),
+            timeindex=pd.date_range("1/1/2012", periods=2, freq="h"),
             timeincrement=[1, 2, 3, 4],
             infer_last_interval=False,
         )
@@ -147,7 +147,7 @@ def test_missing_timeincrement():
 
 def test_overwrite_timeincrement():
     es = solph.EnergySystem(
-        timeindex=pd.date_range("1/1/2012", periods=2, freq="H"),
+        timeindex=pd.date_range("1/1/2012", periods=2, freq="h"),
         infer_last_interval=True,
     )
     assert es.timeincrement[0] == 1
@@ -163,8 +163,8 @@ def test_model_timeincrement_list():
 
 def test_nonequ_inconsistent_timeindex():
     # with pytest.raises(IndexError):
-    timeindex_one = pd.date_range("1/1/2019", periods=1, freq="H")
-    timeindex_hourly = pd.date_range("1/1/2019", periods=3, freq="H")
+    timeindex_one = pd.date_range("1/1/2019", periods=1, freq="h")
+    timeindex_hourly = pd.date_range("1/1/2019", periods=3, freq="h")
     timeindex_45mins = pd.date_range("1/1/2019", periods=2, freq="45min")
     timeindex1 = timeindex_one.append(timeindex_hourly)
     timeindex2 = timeindex_hourly.append([timeindex_45mins])
